[script] fix output when no arguments are specified (#5551)
When no arguments are provided to script/test, the script prints $1:
unbound variable and exits without printing the usage
information. This is due to the set -u at the top.

This commit has a small fix so that this usage is presented with the
usage information. Also, it prints an error when it detect an unknown
command.