[thread-cert] turn off OTBR NAT64 in 1.2 Backbone tests (#5652)
This commit fixes random fails in Thread 1.2 Backbone tests that uses
OTBR Docker.

The fails are caused by failing to start tayga service when there are
many OTBR Dockers running.  This commit turns off NAT64 feature of
OTBR in 1.2 Backbone tests to avoid this issue since NAT64 is not
useful for Backbone tests.