[nrf52x] allow users to implement power setting functions
When users add a FEM after the radio chip, the actual output power is
determined by both the radio chip and FEM. The radio driver only provide
a function to set the radio output power. This PR adds Macros and functions
to allow users to implement the power setting functions by themselves.