[low-power] fix csl simulation (#5785)
On nRF platform, otPlatRadioGetNow always gets the LSB part of an
uint64_t time. However, on simulation platform, this is not the
case. So in

  if (otPlatRadioGetNow(&GetInstance()) <
    mTransmitFrame.mInfo.mTxInfo.mTxDelayBaseTime +
      mTransmitFrame.mInfo.mTxInfo.mTxDelay)

we don't get the result expected. otPlatRadioGetNow(&GetInstance()) is
a very big number on simulation platform, and mTxDelayBaseTime is the
LSB part of the rx time.