[dns-header] fix 'ParseName()' when header offset is non-zero
This commit fixes an issue with `Dns::Name::ParseName()`. This method
intentionally does not get the header offset (which is used for
parsing compressed DNS names) since its implementation does not need
to read the entire name and can stop as soon as we reach the end of
the current segment of the name. This commit changes `GetNextLabel()`
method adding a new input parameter `aStopOnEndOffsetSet` to stop as
soon as `mNameEndOffset` is found and set on the `Iterator`. This new
parameter is then used by `ParseName()` ensuring that we stop
immediately when we find the end offset and do not continue reading a
next label (jumping to an incorrect offset and potentially reading a
bad label length value and reporting it as a parse error).

The unit test `test_dns` is also updated to catch and cover this
situation by populating the guard region in the start of the message
with `0x80 + index` values which are invalid label length values
(length value's first two bits must be 00 or 11).