[settings] add APIs to securely store critical settings
All the settings are written into a local file on the posix platform.
It is easy for others to get the critical infomation, such as master key,
from the local file. This PR adds new APIs `otPlatSecureSettingsXxx` to
allow the vendor to store the critical settings to a secure area.