[dataset] change updater to use DatasetManager (#6227)
This commit changes the dataset updater implementation to leverage
MeshCoP::DatasetManager, which already provides a mechanism to
communicate Pending Dataset updates with the Leader. The
DatasetManager implementation ensures that only one CoAP transaction
is outstanding at a time.

This commit also ensures that any unspecified values are copied from
the Active Dataset.