Fix a bug with Thread v1.2 RCP failing to MAC encrypt outgoing packets properly.
In Thread v1.2, we enable OPENTHREAD_CONFIG_MAC_SOFTWARE_TX_SECURITY_ENABLE on the RCP, since we have to modify the MAC-layer packet for outgoing transmissions in some cases (adding IEs for enhanced ACKs, including CSL sync info etc).  For this purpose, we define otPlatRadioSetMacKey and otPlatRadioSetMacFrameCounter to store the latest security material for AES-CCM encryption.

However, the bug was that this info was not being inserted and transmit security not processed on the outgoing packet, since the mac_fr expects calls to otMacFrameSetKeyId, otMacFrameSetFrameCounter, and otMacFrameProcessTransmitAesCcm from the user.  Included this piece of code to fix the Thread v1.2 RCP bug.

Other changes:
1. Moved a few pieces of helper code to better organize various 1.2 sections in radio.c, and fix some compilation errors.
2. Added some more clarifying comments in some places.
3. Re-enabled some CSL code which was commented out by accident.