[ip6] add operator `<` overload for 'Ip6::Prefix' comparison (#6274)
This commit adds an overload of operator `<` in `Ip6::Prefix` class. A
prefix with shorter length is considered smaller than the one with
longer length. If the prefix lengths are equal, then the prefix bytes
are compared directly to determine the smaller prefix. This commit
then changes `RoutingManager` to use the newly added operator `<`
overload.

Unit test `test_ip6_address.cpp` is updated to cover the behavior of
the new operator.